requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
Django==3.2.6
django-bootstrap4==3.0.1
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
gunicorn==20.1.0
Markdown==3.3.4
PyJWT==2.1.0
pykerberos==1.2.1
pymongo==3.12.0
python-snappy==0.6.0
pytz==2021.1
PyYAML==5.4.1
soupsieve==2.2.1
sqlparse==0.4.1
